+++ src/main/java/org/jcd2052/api/exceptionhandler/exceptions/GameGenreNotFoundException.java
@@ -1,7 +1,7 @@
 package org.jcd2052.api.exceptionhandler.exceptions;
 
 public class GameGenreNotFoundException extends RuntimeException {
-    public GameGenreNotFoundException(int genreId) {
+    public GameGenreNotFoundException(long genreId) {
         super("Couldn't find Game Genre with id " + genreId);
     }
 }
