+++ src/main/java/org/jcd2052/api/exceptionhandler/exceptions/GameReviewNotFoundException.java
@@ -1,7 +1,7 @@
 package org.jcd2052.api.exceptionhandler.exceptions;
 
 public class GameReviewNotFoundException extends RuntimeException {
-    public GameReviewNotFoundException(int reviewId) {
+    public GameReviewNotFoundException(long reviewId) {
         super(String.format("Review with id %d has been not found.", reviewId));
     }
 }
