+++ src/main/java/org/jcd2052/api/exceptionhandler/exceptions/PlatformNotFoundException.java
@@ -1,7 +1,7 @@
 package org.jcd2052.api.exceptionhandler.exceptions;
 
 public class PlatformNotFoundException extends RuntimeException {
-    public PlatformNotFoundException(int platformId) {
+    public PlatformNotFoundException(long platformId) {
         super("Couldn't find platform with id " + platformId);
     }
 }
